master - update 0409 11:34

+++ YUK/src/main/java/com/itwillbs/controller/AjaxController.java
@@ -3,6 +3,7 @@ package com.itwillbs.controller;
 import java.util.List;
 
 import javax.inject.Inject;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.itwillbs.domain.ProductDTO;
 import com.itwillbs.domain.ProductionDTO;
@@ -52,28 +54,37 @@ public class AjaxController {
 	@PostMapping("/production/updatePer")
 	public ResponseEntity<?> updatePer(@RequestBody ProductionDTO productionDTO, Authentication authentication) {
 	    System.out.println("ProductionController updatePer()");
-	    System.out.println(productionDTO);
+	    
 		String username = authentication.getName();
 		productionDTO.setName(username);	
-
-	    productionService.updatePer(productionDTO);
-	    
-	    // 성공 응답 보내기
-	    // 단순 문자열 "success"를 JSON 형식으로 반환하는 예시
-	    return ResponseEntity.ok().body("{\"message\":\"success\"}");
+		
+		ProductionDTO productionDTO2 = productionService.getPib(productionDTO);
+		if(productionDTO2.getPibStatus() == 0) {
+			productionService.updatePer(productionDTO);
+			// 성공 응답 보내기
+			// 단순 문자열 "success"를 JSON 형식으로 반환하는 예시
+			return ResponseEntity.ok().body("{\"message\":\"success\"}");
+		} else {
+			return ResponseEntity.badRequest().body("{\"message\":\"Operation not allowed\"}");
+		}
+		
 	}
 	
 	// 실적 삭제
 	@PostMapping("/production/perDeletePro")
 	public ResponseEntity<?> perDeletePro(@RequestBody ProductionDTO productionDTO) {
 	    System.out.println("ProductionController perDeletePro()");
-	    System.out.println(productionDTO);
-
-	    productionService.deletePer(productionDTO);
 	    
-	    // 성공 응답 보내기
-	    // 단순 문자열 "success"를 JSON 형식으로 반환하는 예시
-	    return ResponseEntity.ok().body("{\"message\":\"success\"}");
+		ProductionDTO productionDTO2 = productionService.getPib(productionDTO);
+		if(productionDTO2.getPibStatus() == 0) {
+			productionService.deletePer(productionDTO);
+			// 성공 응답 보내기
+			// 단순 문자열 "success"를 JSON 형식으로 반환하는 예시
+			return ResponseEntity.ok().body("{\"message\":\"success\"}");
+		} else {
+			return ResponseEntity.badRequest().body("{\"message\":\"Operation not allowed\"}");
+		}
+			
 	}
 
 	// 소요량 가져오기
@@ -88,7 +99,20 @@ public class AjaxController {
 	    return ResponseEntity.ok().body(reqList);
 	}
 
-	
+	@PostMapping("/production/lineDeletePro")
+	public ResponseEntity<?> lineUpdatelineDeleteProPro(@RequestBody ProductionDTO productionDTO, HttpSession session, RedirectAttributes redirectAttributes) {
+		System.out.println("ProductionController lineDeletePro()");
+		System.out.println(productionDTO);
+		
+		ProductionDTO productionDTO2 = productionService.getLine(productionDTO);
+		if(productionDTO2.getLineStatus() != 1) {
+			productionService.deleteLine(productionDTO);
+			return ResponseEntity.ok().body("{\"message\":\"success\"}");
+		} else {
+	        redirectAttributes.addFlashAttribute("errorMessage", "가동중인 라인은 수정/삭제가 불가능합니다.");
+	        return ResponseEntity.badRequest().body("{\"message\":\"Operation not allowed\"}");
+		}
+	}
 	
 	
 }
